footer in its own react
