+++ src/App/app-main.js
@@ -6,6 +6,7 @@ import { connect } from 'react-redux';
 import authUtils from './authUtils';
 import mapStoreToProps from '../redux/mapStoreToProps';
 import appMainUtils from './appMainUtils';
+import Footer from './Footer';
 
 export class AppTemplate extends Component {
   constructor(props) {
@@ -17,7 +18,7 @@ export class AppTemplate extends Component {
     this.handleKeyMenu = this.handleKeyMenu.bind(this);
     this.toggleMobileMenu = this.toggleMobileMenu.bind(this);
     this.changeNav = this.changeNav.bind(this);
-    this.footerLinks = this.footerLinks.bind(this);
+
     this.navLinks = this.navLinks.bind(this);
     this.responseGoogleLogin = this.responseGoogleLogin.bind(this);
     this.responseGoogleLogout = this.responseGoogleLogout.bind(this);
@@ -165,33 +166,6 @@ export class AppTemplate extends Component {
     );
   }
 
-  footerLinks() { // eslint-disable-line class-methods-use-this
-    const color = '#c09580';
-    const links = [
-      { href: 'https://github.com/WebJamApps', name: 'github' },
-      { href: 'https://www.linkedin.com/company/webjam/', name: 'linkedin' },
-      { href: 'https://www.instagram.com/joshua.v.sherman/', name: 'instagram' },
-      { href: 'https://twitter.com/WebJamLLC', name: 'twitter' },
-      { href: 'https://www.facebook.com/WebJamLLC/', name: 'facebook' },
-    ];
-    return (
-      <div style={{ textAlign: 'center', padding: '6px' }}>
-        {
-          links.map((link) => (
-            <a key={Math.random().toString()} target="_blank" rel="noopener noreferrer" style={{ color, paddingRight: '5px' }} href={link.href}>
-              <span><i className={`fab fa-${link.name}`} /></span>
-            </a>
-          ))
-        }
-        <p style={{ color: 'white', fontSize: '9pt', marginBottom: 0 }}>
-          Powered by
-          {' '}
-          <a className="wjllc" target="_blank" rel="noopener noreferrer" href="https://www.web-jam.com">Web Jam LLC</a>
-        </p>
-      </div>
-    );
-  }
-
   headerSection() {
     return (
       <div id="header" className={`material-header ${this.currentStyles.headerClass}`}>
@@ -227,12 +201,13 @@ export class AppTemplate extends Component {
             {this.headerSection()}
             <div style={{ width: 'auto' }} id="contentBlock" className="content-block">
               { this.children }
-              <div id="wjfooter" className="footer" style={{ backgroundColor: '#565656', paddingTop: '20px', paddingBottom: '40px' }}>
-                { this.footerLinks() }
-              </div>
+              <Footer />
             </div>
+
           </div>
+
         </div>
+
       </div>
     );
   }
